Refactor solver preference searching.
This resolves a couple bugs in the solver's preference search. First,
multiple ordered constraints (i.e. Dependency) on a single
Installable are searched properly now. Second, the search will make a
selection for each dependency constraint in the solution, whereas
before it would bail as soon as it's known to be SAT.

Searching has also been refactored to be tested independently from
solving, with a fake of S.